删除前先判断是否存在

+++ iclientpy/conda/build.py
@@ -2,14 +2,16 @@
 import os
 import sys
 import shutil
+import pathlib
 conda_build_dir = os.path.abspath(os.path.join(sys.argv[0], os.pardir))
 output_dir = os.path.join(conda_build_dir, 'output')
-shutil.rmtree(output_dir)
+if pathlib.Path(output_dir).exists():
+    shutil.rmtree(output_dir)
 get_ipython().system('conda build -c https://mirrors.tuna.tsinghua.edu.cn/anaconda/cloud/conda-forge/ --output-folder {output_dir} iclientpy')
-import pathlib
 cur_platform_name = [name for name in os.listdir(output_dir) if name != 'noarch'][0]
 channel_dir = os.path.join(conda_build_dir, 'channel')
-shutil.rmtree(channel_dir)
+if pathlib.Path(channel_dir).exists():
+    shutil.rmtree(channel_dir)
 cur_platform_dir = os.path.join(channel_dir, cur_platform_name)
 pathlib.Path(channel_dir).mkdir(parents=True, exist_ok=True)
 shutil.copytree(os.path.join(output_dir, cur_platform_name), cur_platform_dir)
@@ -20,7 +22,8 @@ get_ipython().system('conda index {channel_dir}')
 noarch_dir = os.path.join(channel_dir, 'noarch')
 get_ipython().system('conda index {noarch_dir}')
 tar_file_path = os.path.join(conda_build_dir, 'iclientpy-conda-package.tar')
-os.remove(tar_file_path)
+if pathlib.Path(tar_file_path).exists():
+    os.remove(tar_file_path)
 import tarfile
 with tarfile.open(tar_file_path, 'w') as zipf:
     for root,dirs,files in os.walk(channel_dir):
